+++ lib/src/bloc/bloc_provider.widget.dart
@@ -25,9 +25,10 @@ class BLoCProvider<T extends BLoC> extends StatefulWidget {
   _BLoCProviderState<T> createState() => _BLoCProviderState<T>();
 
   static T of<T extends BLoC>(BuildContext context) {
-    final type = _typeOf<BLoCProvider<T>>();
-    BLoCProvider<T> provider = context.ancestorWidgetOfExactType(type);
-    return provider.bloc;
+    final type = _typeOf<_BLoCProviderInherited<T>>();
+    _BLoCProviderInherited<T> provider =
+        context.ancestorInheritedElementForWidgetOfExactType(type)?.widget;
+    return provider?.bloc;
   }
 
   static Type _typeOf<T>() => T;
@@ -47,7 +48,12 @@ class _BLoCProviderState<T extends BLoC> extends State<BLoCProvider<T>> {
   }
 
   @override
-  Widget build(BuildContext context) => widget.child;
+  Widget build(BuildContext context) {
+    return _BLoCProviderInherited(
+      child: widget.child,
+      bloc: widget.bloc,
+    );
+  }
 
   @override
   void reassemble() {
@@ -64,4 +70,17 @@ class _BLoCProviderState<T extends BLoC> extends State<BLoCProvider<T>> {
     widget.bloc.close();
     super.dispose();
   }
+}
+
+class _BLoCProviderInherited<T> extends InheritedWidget {
+  _BLoCProviderInherited({
+    Key key,
+    @required Widget child,
+    @required this.bloc,
+  }) : super(key: key, child: child);
+
+  final T bloc;
+
+  @override
+  bool updateShouldNotify(_BLoCProviderInherited oldWidget) => false;
 }