+++ lib/src/ui/preferred_async_builder.widget.dart
@@ -47,7 +47,10 @@ class PreferredFutureBuilder<T> extends StatelessWidget {
       future: future,
       builder: (ctx, snapshot) {
         if (snapshot.hasError) {
-          L.p(snapshot.error);
+          L.p('PreferredFutureBuilder出现错误: ${snapshot.error}');
+          if (snapshot.error is Error) {
+            L.p((snapshot.error as Error).stackTrace);
+          }
           if (errorPlaceholderBuilder != null) {
             return errorPlaceholderBuilder(context, snapshot.error);
           } else {
@@ -111,7 +114,10 @@ class PreferredStreamBuilder<T> extends StatelessWidget {
       stream: stream,
       builder: (ctx, snapshot) {
         if (snapshot.hasError) {
-          L.p(snapshot.error);
+          L.p('PreferredStreamBuilder出现错误: ${snapshot.error}');
+          if (snapshot.error is Error) {
+            L.p((snapshot.error as Error).stackTrace);
+          }
           if (errorPlaceholderBuilder != null) {
             return errorPlaceholderBuilder(context, snapshot.error);
           } else {