+++ lib/src/utils/functions.dart
@@ -1,1 +1,2 @@
 final notEqual = (prev, next) => prev != next;
+final notNull = (data) => data != null;